restruct fused_triton file to avoid too long file
