+++ python/sglang/srt/layers/moe/fused_moe_triton/__init__.py
@@ -2,10 +2,9 @@ from contextlib import contextmanager
 from typing import Any, Dict, Optional
 
 import sglang.srt.layers.moe.fused_moe_triton.fused_moe  # noqa
-from sglang.srt.layers.moe.fused_moe_triton.fused_moe import (
-    fused_experts,
-    get_config_file_name,
-)
+from sglang.srt.layers.moe.fused_moe_triton.fused_moe import fused_experts
+from sglang.srt.layers.moe.fused_moe_triton.fused_moe_triton_config import get_config_file_name
+
 from sglang.srt.layers.moe.fused_moe_triton.layer import (
     FusedMoE,
     FusedMoEMethodBase,
